requirements.txt
torch
torchvision
numpy
pandas
Pillow
scikit-learn
matplotlib
tqdm
mlflow